finish splitting cli.tsx into cli/ and ui/ modules
