+++ src/ui/markdown.tsx
@@ -1,0 +1,189 @@
+import React from "react";
+import { Box, Text } from "ink";
+import { marked, type Token, type Tokens } from "marked";
+import { stripHtmlTags } from "../utils.js";
+
+/**
+ * Renders a Markdown string as Ink text: lexes it with `marked` and maps each
+ * block and inline token to terminal output (headings, lists, code, tables,
+ * emphasis), stripping any HTML down to plain text.
+ */
+export function MarkdownText({ markdown }: { markdown: string }) {
+  const tokens = marked.lexer(markdown, {
+    async: false,
+    gfm: true,
+  });
+
+  return (
+    <Box flexDirection="column">
+      {tokens.map((token, index) => (
+        <MarkdownBlock
+          index={index}
+          key={`${token.type}-${index}`}
+          token={token}
+        />
+      ))}
+    </Box>
+  );
+}
+
+function MarkdownBlock({ index, token }: { index: number; token: Token }) {
+  if (token.type === "space" || token.type === "def" || token.type === "hr") {
+    return null;
+  }
+
+  if (token.type === "paragraph") {
+    return (
+      <Text wrap="wrap">
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "heading") {
+    return (
+      <Text wrap="wrap">
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "list") {
+    return (
+      <Box flexDirection="column">
+        {(token as Tokens.List).items.map((item, itemIndex) => (
+          <Text key={`${index}-${itemIndex}`} wrap="wrap">
+            <Text color="gray">
+              {(token as Tokens.List).ordered
+                ? `${Number((token as Tokens.List).start || 1) + itemIndex}. `
+                : "- "}
+            </Text>
+            <InlineMarkdown tokens={getTokenChildren(item)} />
+          </Text>
+        ))}
+      </Box>
+    );
+  }
+
+  if (token.type === "code") {
+    return <Text color="gray">{token.text}</Text>;
+  }
+
+  if (token.type === "blockquote") {
+    return (
+      <Text wrap="wrap">
+        <Text color="gray">| </Text>
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "table") {
+    return <Text color="gray">{renderPlainTable(token as Tokens.Table)}</Text>;
+  }
+
+  if (token.type === "html") {
+    return <Text wrap="wrap">{renderHtmlToken(token)}</Text>;
+  }
+
+  if (token.type === "text") {
+    return (
+      <Text wrap="wrap">
+        <InlineMarkdown tokens={token.tokens ?? [token]} />
+      </Text>
+    );
+  }
+
+  return <Text wrap="wrap">{token.raw}</Text>;
+}
+
+function InlineMarkdown({ tokens }: { tokens: Token[] }) {
+  return (
+    <>
+      {tokens.map((token, index) => (
+        <InlineMarkdownToken key={`${token.type}-${index}`} token={token} />
+      ))}
+    </>
+  );
+}
+
+function InlineMarkdownToken({ token }: { token: Token }) {
+  if (token.type === "text" || token.type === "escape") {
+    return <>{token.text}</>;
+  }
+
+  if (token.type === "strong") {
+    return (
+      <Text bold>
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "em") {
+    return (
+      <Text italic>
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "link") {
+    return (
+      <Text underline>
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "codespan") {
+    return <Text color="gray">{token.text}</Text>;
+  }
+
+  if (token.type === "br") {
+    return <>{"\n"}</>;
+  }
+
+  if (token.type === "del") {
+    return (
+      <Text strikethrough>
+        <InlineMarkdown tokens={getTokenChildren(token)} />
+      </Text>
+    );
+  }
+
+  if (token.type === "html") {
+    return <>{renderHtmlToken(token)}</>;
+  }
+
+  if ("tokens" in token && Array.isArray(token.tokens)) {
+    return <InlineMarkdown tokens={token.tokens} />;
+  }
+
+  return <>{token.raw}</>;
+}
+
+function getTokenChildren(token: Token): Token[] {
+  return "tokens" in token && Array.isArray(token.tokens) ? token.tokens : [];
+}
+
+function renderPlainTable(token: Tokens.Table): string {
+  const header = token.header.map((cell) => cell.text).join(" | ");
+  const rows = token.rows.map((row) =>
+    row.map((cell) => cell.text).join(" | "),
+  );
+
+  return [header, ...rows].filter(Boolean).join("\n");
+}
+
+function renderHtmlToken(token: Token): React.ReactNode {
+  const text =
+    "text" in token && typeof token.text === "string" ? token.text : token.raw;
+  const underlineMatch = text.match(/^<u>(.*)<\/u>$/isu);
+
+  if (underlineMatch) {
+    return <Text underline>{underlineMatch[1]}</Text>;
+  }
+
+  return stripHtmlTags(text);
+}
